Add functions necessary for updates to testlib
This commit adds a couple of new helper functions to the testlib
that will be useful when creating tests for updates and verify.
These are the functions added by this commit:
- create_version
- update_bundle

This commit also adds a couple new assertions that can be used to
compare files:
- assert_files_equal
- assert_files_not_equal

Finally, this commmit fixes a common bug present in many functions
in the test lib, where thr $ was missing from many variables.