Updating tests that use a fake upstream web server
Recently the test library got some functions added that allow users to
define a different default upstream server for the test environment.
This was aimed to simplify the way we are writing the tests and to make
them more similar to the way a production system works.

This commit leverages from those new functions to simplify tests that
use a web server.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>