Balance tests in jobs in Travis
This commit includes two more improvements for the swupd build process:

- It balances the tests run in each one of the four test jobs so they
take aproximatelly the same time. Before tests were being balanced per
test directory, with this commit tests are now balanced at a test file
level, which provides an even better tunning. In this case we are
splitting the update tests into 2 jobs since they are the most time
consuming.

- Instead of using bats directly for running tests, this commit uses
"make check" to run the tests but it provides a specific list for make
to override the default list so it doesn't run all tests. The advantage
of this is to still enjoy the benefits of the nice output provided by
make check which only shows a list of the tests executed and their status
and it only shows the the test output from the tests that failed.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>