Improve output showing dependencies
When listing the dependencies of a given bundle, swupd uses a tree view
that includes all bundles that depend on the specified bundle, plus all
the bundles that depend on that dependendant bundle, and so on.
Some times for some bundles the list can grow quite large (over 1000
lines) which make the output useless.

This commit fixes the issue by simplifying the list of dependencies if
the list is big so it is easier to read.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>