Refactor remove to delete multiple bundles at once
The bundle-remove process was validating and trying to remove all
provided bundles one by one. This commit refactors this process so the
provided bundles are first validated one by one, but they are removed
all at once. This change in processing removals will make it easier to
implement the removal of a bundle along with its dependencies, which
will be implemented in a future commit.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>