Add --force flag to bundle-remove
When attempting to remove a bundle that is required (included by) by
other bundles, swupd won't allow it. This makes removing such bundles
very difficult because you need to delete them one by one.

This commit adds a "--force" flag that can be used to remove the bundle
along with all the bundles that depend on it.

Closes #891

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>