Consider all bundles in args using bundle-remove
When running bundle-remove with multiple bundles, as in "bundle-remove
bundle1 bundle2" if bundle1 is required by bundle2, bundle1 will fail to
be removed because another installed bundle depends on it (bundle2),
however bundle2 will be removed successfully because no other bundle
depends on it. If the order of the bundles is inverted in the command,
as in "bundle-remove bundle2 bundle1", then both bundles will be removed
successfully since they are analyzed in the order they are provided by
the user. This is a confusing output for the user.

This commit changes the bundle-remove command so if a bundle and all
bundles that depend on it are added to the command regardless of the
order, they will be removed successfully.

Closes #674

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>