Testlib: creating the test env only once
When the global_setup is used we only create the test environment once
and use it for every test in the file. However when test_setup is used,
every test use their own isolated test environment, so we create one
environment per test at the beginning of the test. Since the test
environments are the same for each test, instead of creating an
environment per test we can create only one environment at the beginning
of all tests and create copies for each test in the file. Copying is a
lot faster than regenerating the environment from scratch.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>